Add Storage class and incorporate both TaskList and Storage into Duchess
